# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "e-Test API",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "security": [
        {
           "Bearer": []
        }
    ],
    "consumes": [],
    "produces": [],
    "paths": {
...
        "/api/v1/test/{id}/sections": {
            "get": {
                "tags": ["test"],
                "summary":"Get all sections in a test",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/responses/Section"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/test/{id}/questions": {
            "get": {
                "tags": ["test"],
                "summary":"Get all questions in a test",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/responses/TestQuestions"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/v1/section/": {
            "get": {
                "tags": ["section"],
                "summary": "Get all sections",
                "description": "",
                "parameters": [
                    {
                        "name": "select",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/responses/Section"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": ["section"],
                "summary": "Create section",
                "description": "",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Section"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/responses/Section"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Test": {
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Some example string"
                },
                "timer": {
                    "type": "integer",
                    "default": 0
                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
        "Section": {
            "required": [
                "title",
                "timer",
                "instruction",
                "test"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Some example string"
                },
                "timer": {
                    "type": "integer",
                    "default": 0
                },
                "instruction": {
                    "type": "string",
                    "example": "Some example string"
                },
                "test": {
                    "type": "string",
                    "example": "Test id"
                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
        "Question": {
            "required": [
                "question",
                "correct_answers",
                "section"
            ],
            "properties": {
                "question": {
                    "type": "string",
                    "example": "Some example string"
                },  
                "answers": {
                    "type": "array",
                    "example": ["Some", "example", "string"]
                },
                "correct_answers": {
                    "type": "array",
                    "example": ["Some", "example", "string"]
                },
                "section": {
                    "type": "string",
                    "example": "Section id"
                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
...
    },
    "responses": {
        "Section": {
            "required": [
                "title",
                "timer",
                "instruction",
                "test"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Some example string"
                },
                "timer": {
                    "type": "integer",
                    "default": 0
                },
                "instruction": {
                    "type": "string",
                    "example": "Some example string"
                },
                "test": {
                    "$ref": "#/definitions/Test"                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
...
        "TestQuestions": {
            "required": ["file"],
            "properties": {
                "data": {
                    "type": "array",
                    "example": [
                        {
                            "section": {
                                "$ref": "#/definitions/Section"                
                            },
                            "question": {
                                "$ref": "#/definitions/Question"
                            }
                        }
                    ]
                }
            }
        }
    }
}